Made modular with blueprint
Refractored main.py and used blueprint to refactor the routes and separated the functions into different folders
