+++ pyproject.toml
@@ -15,6 +15,7 @@ dropbox = "^12.0.2"
 werkzeug = "^3.0.3"
 api = "^0.0.7"
 requests = "^2.32.3"
+route = "^0.1.4"
 
 [tool.pyright]
 # https://github.com/microsoft/pyright/blob/main/docs/configuration.md
