+++ replit.nix
@@ -1,5 +1,6 @@
 {pkgs}: {
   deps = [
+    pkgs.python-launcher
     pkgs.nodePackages.prettier
     pkgs.libxcrypt
     pkgs.rustc
