+++ templates/all-applications.html
@@ -1,4 +1,3 @@
-<!-- Import Jinja Macros -->
 {% from 'macros.html' import navbar, call_to_actions as actions, list %}
 
 <!DOCTYPE html>
@@ -6,7 +5,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>All Companies (for testing)</title>
+    <title>Veric</title>
 
     <!-- External CSS -->
     <link rel="stylesheet" href="../static/css/style.css" />
@@ -43,52 +42,44 @@
       };
     </script>
   </head>
-
   <body class="flex flex-col items-center justify-start min-h-screen">
-    <!-- Header -->
     {% set nav_links = [ { "href": "/login", "name": "Login" }, { "href":
     "/sign-up", "name": "Sign up" }, ] %} {% set logo = '<span
       class="text-3xl font-bold text-gray-800"
       >Veric</span
-    >' %} {{ navbar(logo=logo | safe, navlinks=nav_links) }}
+    >' %} {{ navbar(logo=logo|safe, navlinks=nav_links) }}
 
-    <!-- Main Content -->
-    <main
-      class="w-full flex flex-col items-center justify-center p-8 space-y-8 lg:px-12"
-    >
-      <div class="space-y-8 w-full">
+    <main class="w-full h-full">
+      <section class="relative flex">
         <div
-          class="shadow-sm flex flex-col items-start justify-center p-4 space-y-2 border border-gray-300 rounded-lg bg-white w-4/5"
+          class="left-0 flex flex-col w-2/3 bg-white border border-2 absolute"
         >
-          {% set displayed_fields = [
-                { "name": "id", "key": "id", "type": "text", "editable": False },
-                { "name": "Company Name", "editable": True, "key": "name", "type": "text" },
-                { "name": "Company Email", "editable": True, "key": "email", "type": "email" },
-                { "name": "Company Phone Number", "editable": True, "key": "contact", "type": "tel" },
-                { "name": "Logo", "editable": True, "key": "logo", "type": "image" }
-            ] %}
+        {% set displayed_fields = [
+          {"name" : "id", "key" : "id", "type" : "text", "editable" : False },
+          {"name": "Application name", "editable": True, "key": "jobName", "type": "text"},
+          {"name": "Posted by", "editable": False, "key": "companyName", "type": "text"},
+          {"name": "Role", "editable": False, "key": "role", "type": "checkbox"},
+          {"name": "Logo", "editable": False, "key": "logo", "type": "image"},
+        ]%}
 
-          {{ list( 
-            data=data,
-            fields=displayed_fields,
-            is_editable=false,
-            table=table,
-            has_image=True,
-            is_addable=False,
-            is_deletable=False,
-          )}}
+        {{ list( 
+          data=data,
+          fields=displayed_fields,
+          is_editable=false,
+          table=table,
+          has_image=True,
+          is_addable=False,
+          is_deletable=False,
+        )}}
         </div>
-      </div>
+        <div class="right-0 w-1/3 h-full absolute"></div>
+      </section>
     </main>
 
-    <!-- Footer -->
-    <footer
-      class="h-40 flex flex-col items-center justify-center w-full space-y-4"
-    >
-      <img src="../static/images/logo-black.svg" alt="Logo" class="w-16" />
-      <p class="text-center text-gray-700">
-        &copy; 2023 Veric. All rights reserved.
-      </p>
-    </footer>
+    <script
+      src="https://kit.fontawesome.com/68c16b630a.js"
+      crossorigin="anonymous"
+    ></script>
+    <!-- Include Font Awesome -->
   </body>
 </html>
