+++ templates/login.html
@@ -55,7 +55,7 @@
         class="w-[24rem] h-[24rem] mx-auto p-[6rem] border border-gray rounded-2xl shadow-md box-border"
       >
         <h1 class="text-3xl mb-[2rem] text-gray-800 mx-auto">Log In</h1>
-        {{ render_form(url_for('login'), 'POST', [ {'id': 'email', 'name':
+        {{ render_form('/login', 'POST', [ {'id': 'email', 'name':
         'email', 'type': 'email', 'placeholder': 'Email', 'required': True},
         {'id': 'password', 'name': 'password', 'type': 'password',
         'placeholder': 'Password', 'required': True}, ]) }}
